Mainframe_Platforms/Resilience_APAR_Insights/tools/requirements.txt
requests
pandas
ibm-watson-machine-learning
python-dotenv
ibm_watsonx_ai
typing
ibm-watsonx-orchestrate==1.10.0